fix build hang during testing step of freshly built `script`
The `script` utility intercepts all IO on the standard IO streams by
acting as a pseudo terminal. It starts a fresh shell session that will
be recorded. The shell will listen for input on stdin and will happily
wait forever if none ever arrives. This is what happens during some
(more details further down) builds of the control container image,
causing the build to hang.

Fix this by not actually starting a shell session to see whether the
fresh build of `script` is working, but only emitting its version
information. This is a safe operation never expecting any input, and
hence will not hang the build, but just as effective in seeing whether
anything went catastrophically wrong.

As for the details, I am not sure yet what exactly is going wrong. On
one of my machines the verification step works, on another it hangs the
build. The result does not change even with an old build of the control
container. Rolling back Docker versions and rebuilding does not change
outcomes either.

Tracing `script` and the spawned shell on the functional box shows
"something" injecting an EOT/^D into stdin, which the shell interprets
as `exit`. This causes `script` to exit as well, and the build proceeds.
(Hence the fruitless attempt to downgrade Docker, suspecting that
something around handling stdin may have changed.) I suspect this does
not happen on the box going wrong, but so far was missing the time to
set up tracing for this case as well and actually confirm it.

Signed-off-by: Markus Boehme <markubo@amazon.com>